+++ docs/changes.md
@@ -8,6 +8,7 @@ title: Support Change Log
 ### v2.2.4@dev {#v2-2-4}
 
 * Convert multi-dimensional array to single level array when using `Orchestra\Support\Str::replace()` method.
+* Utilize `Illuminate\Support\Arr`.
 
 ### v2.2.3 {#v2-2-3}